+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `loaders-css` - For animated loading indicators
 - Components:
     - `new-project-navigation-modal` - For navigating away to nodes. Or not.
+    - `settings.account.-components.security` - Two-factor authentication panel.
 - Handbook:
     - `new-project-modal` component
     - `new-project-navigation-modal` component